fix(migration): store usd as Decimal(38, 12)
Cost fields use Decimal(38, 12) per the backend ClickHouse conventions; Float64
accumulates rounding error across a month of summed rows.

+++ apps/opik-backend/src/main/resources/liquibase/db-app-analytics/migrations/000121_create_cipx_savings_facts.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS ${ANALYTICS_DB_DATABASE_NAME}.cipx_savings_facts ON C
     causal_link       LowCardinality(String),
 
     alloc             Float64,
-    usd               Float64,
+    usd               Decimal(38, 12),
 
     last_updated_at   DateTime64(6, 'UTC') DEFAULT now64(6)
 )
